Moving jira api and task list manager
